Consolidate `Error` and `Result`
There does not seem to be any reason to keep `Error` and `Result` in both `vcl` and `vsc` modules.

In the future, I think it would be better to rename them, e.g. to `VarnishError` and `VarnishResult`.  This is a common practice now to distinguish them from the built-in ones (lots of people have been complaining that `std::io::Result` and others didn't follow that convention, but it is too late for that now).
