+++ src/lib.rs
@@ -110,6 +110,8 @@ pub mod ffi {
 
 pub mod vsc;
 
+mod error;
+
 pub mod vcl {
     pub mod backend;
     pub mod convert;
@@ -120,45 +122,9 @@ pub mod vcl {
     pub mod vpriv;
     pub mod vsb;
     pub mod ws;
+    pub use super::error::{Error, Result};
 
     pub mod boilerplate;
-
-    /// custom vcl `Error` type
-    ///
-    /// The C errors aren't typed and are just C strings, so we just wrap them into a proper rust
-    /// `Error`
-    pub struct Error {
-        s: String,
-    }
-
-    impl std::fmt::Debug for Error {
-        fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-            std::fmt::Debug::fmt(&self.s, f)
-        }
-    }
-
-    impl std::fmt::Display for Error {
-        fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-            std::fmt::Display::fmt(&self.s, f)
-        }
-    }
-
-    impl std::error::Error for Error {}
-
-    impl From<String> for Error {
-        fn from(s: String) -> Self {
-            Error { s }
-        }
-    }
-
-    impl From<&str> for Error {
-        fn from(s: &str) -> Self {
-            Error { s: s.into() }
-        }
-    }
-
-    /// Shorthand to `std::result::Result<T, Error>`
-    pub type Result<T> = std::result::Result<T, Error>;
 }
 
 /// Automate VTC testing
